requirements.txt
folium==0.14.0
pandas==2.1.0
numpy==1.24.3
geopy==2.4.0
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
webbrowser 